+++ crates/utils/src/color.rs
@@ -2,7 +2,7 @@
 use std::fmt;
 
 /// A r g b a color.
-#[derive(Copy, Clone, Default)]
+#[derive(Copy, Clone, PartialOrd, Default)]
 #[repr(packed)]
 pub struct Color {
     pub data: u32,
@@ -16,11 +16,79 @@ impl Color {
         }
     }
 
+    /// Create a new color from HSV(0.0-360.0, 0.0-1.0, 0.0-1.0)
+    pub fn hsv(h: f64, s: f64, v: f64) -> Self {
+        Self::hsva(h, s, v, 1.0)
+    }
+
+    /// Create a new color from HSL(0.0-360.0, 0.0-1.0, 0.0-1.0)
+    pub fn hsl(h: f64, s: f64, l: f64) -> Self {
+        Self::hsla(h, s, l, 1.0)
+    }
+
     /// Create a new color from RGB and alpha values
     pub const fn rgba(r: u8, g: u8, b: u8, a: u8) -> Self {
         Color {
             data: ((a as u32) << 24) | ((r as u32) << 16) | ((g as u32) << 8) | (b as u32),
         }
+    }
+
+    /// Create a new color from HSV(0.0-360.0, 0.0-1.0, 0.0-1.0) and alpha values(0.0-1.0)
+    pub fn hsva(mut hue: f64, mut saturation: f64, mut value: f64, alpha: f64) -> Self {
+        hue %= 360.0;
+        saturation = saturation.max(0.0).min(1.0);
+        value = value.max(0.0).min(1.0);
+        let hh = hue / 60.0;
+        let idx = hh.floor() as i32;
+        let ff = hh.fract();
+        let chroma = value * (1.0 - saturation);
+        let second_component = value * (1.0 - (saturation * ff));
+        let t = value * (1.0 - (saturation * (1.0 - ff)));
+        let (r, g, b) = match idx {
+            0 => (value, t, chroma),
+            1 => (second_component, value, chroma),
+            2 => (chroma, value, t),
+            3 => (chroma, second_component, value),
+            4 => (t, chroma, value),
+            5 => (value, chroma, second_component),
+            _ => unreachable!(),
+        };
+        Self::rgba(
+            (r * 255.0) as u8,
+            (g * 255.0) as u8,
+            (b * 255.0) as u8,
+            (alpha * 255.0) as u8,
+        )
+    }
+
+    /// Create a new color from HSL(0.0-360.0, 0.0-1.0, 0.0-1.0) and alpha values(0.0-1.0)
+    pub fn hsla(mut hue: f64, mut saturation: f64, mut lightness: f64, alpha: f64) -> Self {
+        hue %= 360.0;
+        saturation = saturation.max(0.0).min(1.0);
+        lightness = lightness.max(0.0).min(1.0);
+        let hh = hue / 60.0;
+        let idx = hh.floor() as i32;
+        let chroma = (1.0 - ((2.0 * lightness) - 1.0).abs()) * saturation;
+        let second_component = chroma * (1.0 - (((idx % 2) as f64) - 1.0).abs());
+        let (mut r, mut g, mut b) = match idx {
+            0 => (chroma, second_component, 0.0),
+            1 => (second_component, chroma, 0.0),
+            2 => (0.0, chroma, second_component),
+            3 => (0.0, second_component, chroma),
+            4 => (second_component, 0.0, chroma),
+            5 => (chroma, 0.0, second_component),
+            _ => unreachable!(),
+        };
+        let adjustment = lightness - chroma / 2.0;
+        r += adjustment;
+        g += adjustment;
+        b += adjustment;
+        Self::rgba(
+            (r.min(1.0) * 255.0) as u8,
+            (g.min(1.0) * 255.0) as u8,
+            (b.min(1.0) * 255.0) as u8,
+            (alpha * 255.0) as u8,
+        )
     }
 
     /// Get the r value
@@ -74,12 +142,28 @@ impl ToString for Color {
 
 impl From<&str> for Color {
     fn from(s: &str) -> Color {
-        if s == "transparent" {
-            return Color::rgba(0, 0, 0, 0);
-        }
-
         let clean_hex = s.trim_start_matches('#');
         match clean_hex.len() {
+            3 | 4 => {
+                let num = match u32::from_str_radix(&clean_hex, 16) {
+                    Ok(x) => x,
+                    Err(_) => 0,
+                };
+
+                let mut blue = (num & 0xF) << 4;
+                let mut green = ((num >> 4) & 0xF) << 4;
+                let mut red = ((num >> 8) & 0xF) << 4;
+                let mut alpha = match clean_hex.len() == 4 {
+                    true => ((num >> 12) & 0xF) << 4,
+                    false => 0xF,
+                };
+                red |= red >> 4;
+                green |= green >> 4;
+                blue |= blue >> 4;
+                alpha |= alpha >> 4;
+
+                Color::rgba(red as u8, green as u8, blue as u8, alpha as u8)
+            }
             6 | 8 => {
                 let mut x = match u32::from_str_radix(&clean_hex, 16) {
                     Ok(x) => x,